# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation using Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{id}": {
            "get": {
                "summary": "Get a category by ID",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A category"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
        "/api/groups/{id}": {
            "get": {
                "summary": "Get a group by ID",
                "operationId": "8205d284770b45579f4f4c21e9f7988b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A group"
                    },
                    "404": {
                        "description": "Group not found"
                    }
                }
            }
        },
...
        "/api/otzivis/{id}": {
            "get": {
                "summary": "Get a review by ID",
                "operationId": "03c02657f9be7e221b451fa662aefbe6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A review"
                    },
                    "404": {
                        "description": "Review not found"
                    }
                }
            }
        },
...
        "/api/post/category": {
            "get": {
                "summary": "Get products by category",
                "operationId": "85b53ed93099a2d7cc216b1790639b0e",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of products"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "summary": "Get a review by ID",
                "operationId": "142b5adb0520cca80cedfe09d9298332",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A review"
                    },
                    "404": {
                        "description": "Review not found"
                    }
                }
            }
        },
...
        "/api/product-categories/{id}": {
            "get": {
                "summary": "Get a review by ID",
                "operationId": "66bf645e3092240cc3f15c5d0e54b99e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A review"
                    },
                    "404": {
                        "description": "Review not found"
                    }
                }
            }
        },
...
        "/api/product/category": {
            "get": {
                "summary": "Get products by category",
                "operationId": "0c0ba0e02625b513d0c7a30a5680068f",
                "parameters": [
                    {
                        "name": "product_category_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of products"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "summary": "Get a review by ID",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A review"
                    },
                    "404": {
                        "description": "Review not found"
                    }
                }
            }
        },
...
    },
...
}